+++ pyproject.toml
@@ -6,7 +6,7 @@ build-backend = "hatchling.build"
 name='talos'
 description='Centre for Population Genomics Variant Prioritisation'
 readme = "README.md"
-version="8.0.2"
+version="8.0.2.1"
 requires-python = ">=3.10,<3.11"
 license-files = ["LICENSE"]
 classifiers=[
@@ -106,6 +106,13 @@ CreateTalosHTML = 'talos.CreateTalosHTML:cli_main'
 # Generate a brief summary of an existing report (variant count, family size, etc.)
 SummariseReport = 'talos.SummariseReport:cli_main'
 
+# ----------------------------------------------------------------------------------------
+# these scripts are custom HudsonAlpha scripts
+# ----------------------------------------------------------------------------------------
+AddRKReturns = 'talos.hudsonalpha.AddRKReturns:cli_main'
+MakeResultTable = 'talos.hudsonalpha.MakeResultTable:cli_main'
+PluckReturnedResults = 'talos.hudsonalpha.PluckReturnedResults:cli_main'
+
 [tool.hatch.build.targets.wheel]
 packages = ["src/talos"]
 
@@ -169,7 +176,7 @@ hail = ["hail"]
 "src/talos/cpg_internal_scripts/talos_stages.py" = ["ARG002"]
 
 [tool.bumpversion]
-current_version = "8.0.2"
+current_version = "8.0.2.1"
 parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
 serialize = ["{major}.{minor}.{patch}"]
 commit = true
