+++ src/talos/hudsonalpha/example_report_generate.sh
@@ -1,3 +1,4 @@
+# Example of how to manually run CreateTalosHTML.py with Singularity
 export TALOS_CONFIG=/cluster/lab/gcooper/hg38/talos/filtering_config/config.toml
 singularity exec -B /cluster -B /scratch docker://harbor.apps.haib.org/gcooperlab/talos:8.0.2.1 python /cluster/home/jlawlor/talos/src/talos/CreateTalosHTML.py \
         --input /cluster/home/jlawlor/talos/runs/outputs/illumina_september/illumina_september_full_report_2025-09-10_20-07.rkresults.json \
